raop/handlers: avoid strncmp overruns with length checks
In raop_handler_get_parameter and raop_handler_set_parameter, the buffers
passed to strncmp are not necessarily null terminated.  This is sometimes
causing a 1-byte overread.  Instead, check to make sure that the buffer is
large enough for the strncmp to succeed.  This also applies to one use of
strstr, but strnstr is a non-standard extension; therefore, convert to
strncmp + length check also.

Signed-off-by: Derrick Lyndon Pallas <derrick@pallas.us>