Fix NTP sync functionality
The existing NTP synchronization has a few issues. It only waits 3
milliseconds for a response from the iOS device, which is too short of a
delay. Also, if it fails to send the NTP request for some reason, or it
fails to receive the NTP response in time, it stops attempting NTP
synchronization altogether.

This commit attempts to address these problems. The NTP response time
has been increased to 300 ms, and if we fail to send or receive an NTP
packet, we will try again next time. The NTP socket is flushed prior to
each request to ensure we don't read an older delayed response.