Allow choosing renderer at build time
After merging the gstreamer renderers from UxPlay, this project needed a
method for choosing which renderer to build into the program. I opted
for supplying a CMake variable describing which renderer you wish to
use. For backwards-compatibility, it defaults to selecting the openmax
renderer that RPiPlay has historically used.

Because the gstreamer renderer doesn't support several of the options
that are supported by the openmax renderer, those options are removed
from the help output in a gstreamer build.