lib: s/http_parser/llhttp/
The NodeJS project replaced http_parser with llhttp, which is safer, faster,
and easier to maintain than its predecessor.  The license is MIT for both
http_parser and llhttp.  This commit makes the switch for this project, but
depends on a small patch [1] to add the RTSP+RAOP verbs and allow the RTSP
protocol type, & will hopefully be pulled upstream.

Generated code based on llhttp v2.1.3 + patch.

[1] https://github.com/nodejs/llhttp/pull/66