Fixup Raspberry Pi h264 stream hack
The h264bitstream library is quite dangerous because it does not check
bounds on any of the parsed arrays.  However, it is quite useful to be able
to inject an optimized max_dec_frame_buffering into h264 streams on OMX
hardware to lower the rendering pipeline latency.

This change mashses up a couple of things to make h264bitstream usage safe.
First, we do not allocate buffers or the h264 structure until we are pretty
sure we can inject.  Second, we do not read the existing frame and modify
it; instead, we generate a brand new frame with the appropraite flags set
and splice it into the bistream.  Third, we do not cut off the preceding
frame, which could have theoretically happened although was hard to produce
in a practical setup.  Fourth, we only allocate a small modification buffer
and if the NAL would break it, which should not happen with the smaller NAL,
we bail out of the injection.

I've been running the change locally and the change does seem to help with
pipeline latency without having to jump through all the parse/unparse hoops.