lib/crypto: fix up OpenSSL AES calls
Previously, we weren't calling Final after Update when decrypting (or
encrypting) via AES.  My assumption is that this is because it was failing
when due to a padding issue that caused Final to bail out.  This change
turns off padding for both encrypt/decrypt and adds the call to final.

We should probably consider having a static context that is reset after the
call rather than allocating/deallocating a new CTX every time.  I'm going to
punt on that for now though.

Issue: FD-/RPiPlay#283